--- Python Modules/VCmagnets/magnetPYD.depend
@@ -1,175 +1,0 @@
-# depslib dependency file v1.0
-1383690117 source:d:\vela\git projects\vela-clara-controllers\python modules\vcmagnets\main.cpp
-	"main.h"
-
-1383690117 d:\vela\git projects\vela-clara-controllers\python modules\vcmagnets\main.h
-	<windows.h>
-
-1476269720 source:d:\vela\git projects\vela-clara-controllers\hardware\controller.cpp
-	"controller.h"
-	<iostream>
-	<sstream>
-
-1476269736 d:\vela\git projects\vela-clara-controllers\hardware\controller.h
-	"interface.h"
-	"baseObject.h"
-	<boost/python.hpp>
-	<boost/python/suite/indexing/vector_indexing_suite.hpp>
-	<boost/python/suite/indexing/map_indexing_suite.hpp>
-
-1476357729 d:\vela\git projects\vela-clara-controllers\hardware\interface.h
-	<cadef.h>
-	<iostream>
-	<map>
-	"baseObject.h"
-	"structs.h"
-
-1476269217 d:\vela\git projects\vela-clara-controllers\utilities\baseobject.h
-	<string>
-	<vector>
-	<sstream>
-	<cmath>
-	<algorithm>
-
-1476091401 d:\vela\git projects\vela-clara-controllers\structs\structs.h
-	<string>
-	<vector>
-	<cadef.h>
-	<boost/preprocessor.hpp>
-
-1476357869 source:d:\vela\git projects\vela-clara-controllers\hardware\interface.cpp
-	"interface.h"
-	<iostream>
-	<sstream>
-
-1476357869 source:d:\vela\git projects\vela-clara-controllers\hardware\magnet\magnetcontroller.cpp
-	"magnetController.h"
-	<fstream>
-	<iostream>
-
-1476359340 d:\vela\git projects\vela-clara-controllers\hardware\magnet\magnetcontroller.h
-	<string>
-	<vector>
-	<map>
-	"magnetInterface.h"
-	"controller.h"
-	<boost/python.hpp>
-
-1476268724 d:\vela\git projects\vela-clara-controllers\hardware\magnet\magnetinterface.h
-	"interface.h"
-	"magnetStructs.h"
-	"structs.h"
-	"magnetConfigReader.h"
-	<vector>
-	<string>
-	<atomic>
-	<map>
-
-1476180742 d:\vela\git projects\vela-clara-controllers\structs\magnetstructs.h
-	"structs.h"
-	<vector>
-	<string>
-	<map>
-	<thread>
-	<cadef.h>
-
-1476269000 d:\vela\git projects\vela-clara-controllers\utilities\magnetconfigreader.h
-	<string>
-	<vector>
-	<map>
-	<iostream>
-	"configReader.h"
-	"magnetStructs.h"
-
-1476269203 d:\vela\git projects\vela-clara-controllers\utilities\configreader.h
-	"configDefinitions.h"
-	"baseObject.h"
-	<string>
-	<vector>
-	<cadef.h>
-
-1476267038 d:\vela\git projects\vela-clara-controllers\utilities\configdefinitions.h
-	<string>
-
-1476358089 source:d:\vela\git projects\vela-clara-controllers\hardware\magnet\magnetinterface.cpp
-	"magnetInterface.h"
-	"dburt.h"
-	"configDefinitions.h"
-	<iostream>
-	<sstream>
-	<chrono>
-	<algorithm>
-
-1476091401 d:\vela\git projects\vela-clara-controllers\utilities\dburt.h
-	<string>
-	<vector>
-	<map>
-	<iostream>
-	"configReader.h"
-	"dburtDefinitions.h"
-	"magnetStructs.h"
-	"magnetInterface.h"
-
-1476091401 d:\vela\git projects\vela-clara-controllers\utilities\dburtdefinitions.h
-	<string>
-
-1476269189 source:d:\vela\git projects\vela-clara-controllers\utilities\baseobject.cpp
-	"baseObject.h"
-	<time.h>
-
-1476091401 source:d:\vela\git projects\vela-clara-controllers\utilities\configreader.cpp
-	"configReader.h"
-	<iostream>
-	<sstream>
-	<time.h>
-	<algorithm>
-	<fstream>
-	"configDefinitions.h"
-	"structs.h"
-
-1476091401 source:d:\vela\git projects\vela-clara-controllers\utilities\dburt.cpp
-	"dburt.h"
-	"dburtDefinitions.h"
-	"magnetStructs.h"
-	<iostream>
-	<vector>
-	<string>
-	<fstream>
-	<sstream>
-	<time.h>
-	<algorithm>
-	<ctype.h>
-
-1476269048 source:d:\vela\git projects\vela-clara-controllers\utilities\magnetconfigreader.cpp
-	"magnetConfigReader.h"
-	"configDefinitions.h"
-	"RFGunStructs.h"
-	<iostream>
-	<vector>
-	<string>
-	<fstream>
-	<sstream>
-	<time.h>
-	<algorithm>
-	<ctype.h>
-
-1476091401 d:\vela\git projects\vela-clara-controllers\structs\rfgunstructs.h
-	"structs.h"
-	<vector>
-	<string>
-	<map>
-	<cadef.h>
-
-1476272314 source:d:\vela\git projects\vela-clara-controllers\python modules\vcmagnets\src\vcmagnets.cpp
-	"VCmagnets.h"
-	<iostream>
-
-1476365364 d:\vela\git projects\vela-clara-controllers\python modules\vcmagnets\include\vcmagnets.h
-	<string>
-	"magnetController.h"
-	<boost/python/detail/wrap_python.hpp>
-	<boost/python.hpp>
-	<boost/python/suite/indexing/vector_indexing_suite.hpp>
-	<boost/python/suite/indexing/map_indexing_suite.hpp>
-	<boost/python/return_value_policy.hpp>
-
