+++ Python Modules/magnets/magnetPYD.cbp
